[CMS-144] New API JSON Request for GBP currency
This commit changes the JSON API endpoint we use to fetch the GBP
currency. It also submits database changes to the 'diverse_instellingen'
table for the currency field that stores the GBP. DECIMAL(10, 10) was
giving rounding errors, because the currency is usually in 4 decimal
points accurate and the database was expecting 10 decimal accurracy.